Fix potential memory leak when `curl_slist_append` fails (#13)
`curl_slist_append()` returns null on failure, and it won't free the
original list when this happens. Overwriting `self.headers` directly
might leak the original list.

Reasons for the failure are not documented. But the function can only
fail on allocation failure according to the source code. So I chose
`OutOfMemory`.